Controllerの責務分離：画面用とAPI用を別クラスへ分割
StudentControllerを@Controllerに戻し、画面返却のみを担当
StudentApiControllerを新規作成し、JSON返却API（/api）を担当
リソース指向に基づき、URLプレフィックスに /api を追加
