+++ src/main/java/raisetech/studentmanagement/controller/StudentApiController.java
@@ -1,0 +1,39 @@
+package raisetech.studentmanagement.controller;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+import raisetech.studentmanagement.controller.converter.StudentConverter;
+import raisetech.studentmanagement.data.Student;
+import raisetech.studentmanagement.data.StudentsCourses;
+import raisetech.studentmanagement.domain.StudentDetail;
+import raisetech.studentmanagement.service.StudentService;
+
+import java.util.List;
+
+@RestController
+@RequestMapping("/api") // ← 先生が言ってた「プレフィックス（前につける文字）」だよ！
+public class StudentApiController {
+    private StudentService service;
+    private StudentConverter converter;
+
+    @Autowired
+    public StudentApiController(StudentService service, StudentConverter converter) {
+        this.service = service;
+        this.converter = converter;
+    }
+
+    @GetMapping("/studentList")
+    public List<StudentDetail> getStudentList() {
+        List<Student> students = service.searchStudentList();
+        List<StudentsCourses> studentsCourses = service.searchStudentCoursesList();
+
+        return converter.convertStudentDetails(students, studentsCourses);
+    }
+
+    @PostMapping("/updateStudent")
+    public ResponseEntity<String> updateStudent(@RequestBody StudentDetail studentDetail) {
+        service.updateStudent(studentDetail);
+        return ResponseEntity.ok("更新処理が成功しました。");
+    }
+}
