+++ src/main/java/raisetech/studentmanagement/controller/StudentController.java
@@ -1,19 +1,19 @@
 package raisetech.studentmanagement.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.ResponseEntity;
+import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.ModelAttribute;
+import org.springframework.web.bind.annotation.PostMapping;
 import raisetech.studentmanagement.controller.converter.StudentConverter;
 import raisetech.studentmanagement.data.Student;
 import raisetech.studentmanagement.data.StudentsCourses;
 import raisetech.studentmanagement.domain.StudentDetail;
 import raisetech.studentmanagement.service.StudentService;
 
-import java.util.List;
-
-@RestController
+@Controller
 public class StudentController {
     private StudentService service;
     private StudentConverter converter;
@@ -22,15 +22,6 @@ public class StudentController {
     public StudentController(StudentService service, StudentConverter converter) {
         this.service = service;
         this.converter = converter;
-    }
-
-    @GetMapping("/studentList")
-    public List<StudentDetail> getStudentList() {
-        List<Student> students = service.searchStudentList();
-        List<StudentsCourses> studentsCourses = service.searchStudentCoursesList();
-
-        return converter.convertStudentDetails(students, studentsCourses);
-
     }
 
     @GetMapping("/newStudent")
@@ -56,16 +47,9 @@ public class StudentController {
 
     @GetMapping("/updateStudent")
     public String updateStudent(Student student, Model model) {
-
         StudentDetail studentDetail = service.searchStudent(student.getId());
         studentDetail.getStudentsCourses().add(new StudentsCourses());
         model.addAttribute("studentDetail", studentDetail);
         return "updateStudent";
-    }
-
-    @PostMapping("/updateStudent")
-    public ResponseEntity<String> updateStudent(@RequestBody StudentDetail studentDetail) {
-        service.updateStudent(studentDetail);
-        return ResponseEntity.ok("更新処理が成功しました。");
     }
 }
